+++ src/constants/Command.ts
@@ -21,6 +21,7 @@ export const COMMAND = {
   stepMoveDown: `${EXTENSION_NAME}.stepMoveDown`,
   viewStep: `${EXTENSION_NAME}.viewStep`,
   createSnapshot: `${EXTENSION_NAME}.createSnapshot`,
+  createPatch: `${EXTENSION_NAME}.createPatch`,
   // Output
   showOutputChannel: `${EXTENSION_NAME}.showOutputChannel`,
   // Panel view